Use parameters file mapping env

+++ scripts/start.sh
@@ -59,21 +59,25 @@ if [ ! -d "/var/www/html/.git" ]; then
 fi
 
 # Lets get the parameters configs
-echo "Getting parameters"
-
-if [ -d /parameters ];
+if [ "${PARAMETERS_FILE}" != "" ];
 then
-    rm -rf /parameters
-fi
-mkdir -p /parameters
-cd /parameters
-git archive --remote=git@bitbucket.org:krugercorp/production-server-config.git master customers/roles/pltest-app/files/parameters/ | tar xvf -
-mv /parameters/customers/roles/pltest-app/files/parameters/*.yml /parameters/
+    echo "Getting parameters"
 
-if [ -f /parameters/${CONSUL_APPLICATION}.yml ];
-then
-    echo "Found parameters file for ${CONSUL_APPLICATION}"
-    cp /parameters/${CONSUL_APPLICATION}.yml /var/www/html/app/config/parameters.yml
+    if [ -d /parameters ];
+    then
+        rm -rf /parameters
+    fi
+
+    mkdir -p /parameters
+    cd /parameters
+    git archive --remote=git@bitbucket.org:krugercorp/production-server-config.git master customers/roles/pltest-app/files/parameters/ | tar xvf -
+    mv /parameters/customers/roles/pltest-app/files/parameters/*.yml /parameters/
+
+    if [ -f /parameters/${PARAMETERS_FILE} ];
+    then
+        echo "Found parameters file for ${PARAMETERS_FILE}"
+        cp /parameters/${PARAMETERS_FILE} /var/www/html/app/config/parameters.yml
+    fi
 fi
 
 # Always chown webroot for better mounting
